Prep to move DoProteinSearchAction
