--- ms2/src/org/labkey/ms2/query/PeptideFilter.java
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2008-2010 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.labkey.ms2.query;
-
-import org.labkey.api.view.ViewContext;
-
-/**
- * User: jeckels
- * Date: Dec 3, 2008
- */
-public interface PeptideFilter
-{
-    public String getPeptideCustomViewName(ViewContext viewContext);
-}
