+++ protein/api-src/org/labkey/api/protein/ProteinService.java
@@ -94,5 +94,7 @@ public interface ProteinService
 
     ActionURL getPeptideSearchUrl(Container c);
 
+    ActionURL getProteinSearchUrl(Container c);
+
     ActionURL getPeptideSearchUrl(Container c, String sequence);
 }
