+++ protein/api-src/org/labkey/api/protein/search/ProteinSearchWebPart.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2007-2012 LabKey Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.labkey.api.protein.search;
+
+import org.labkey.api.view.JspView;
+
+public class ProteinSearchWebPart extends JspView<ProteinSearchBean>
+{
+    public static final String NAME = "Protein Search";
+
+    public ProteinSearchWebPart(boolean horizontal, ProbabilityProteinSearchForm form)
+    {
+        super(horizontal ? "/org/labkey/ms2/search/searchProteins.jsp" : "/org/labkey/ms2/search/searchProteinsNarrow.jsp");
+        setTitle(NAME);
+        setModelBean(new ProteinSearchBean(horizontal, form));
+    }
+}
